Configure emacs apps with elite_emacs_app lwrp
Create this LWRP to manage and configure emacs apps.
Some apps requires some configuration, this lwrp is designed to manage
how which app is configured. Sensitive data should not be configured in
this manner, such as credentials / tokens because generated files are
plain text. Use another way.
For now only elfeed take a list of feeds as configuration.