Deploy conkyrc files from data bags or attributes
Added elite_conky_rc lwrp to deploy conkyrc files
This resource receive the rc `config` and the rc `text` to generate
the rc file. The `config` attribute is a hash with conky config, and
`text` an array. Each elements on the text array correspond to the
printed line and is a hash: `{ offset: text }`

See the spec's diff for an example.

Moved user config [conky][config] to [config][global_config]. This is
the global configuration and content will be merged with rc config.

Modified the elite_conky lwrp:
 - Add user's global_config attribute
 - Add configs attribute which contains rc files to deploy

Updated conky dzen2 config